requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform
aniso8601==9.0.1
blinker==1.6.2
click==8.1.6
Flask==2.3.2
Flask-RESTful==0.3.10
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.1
pytz==2023.3
schedule==1.2.0
six==1.16.0
Werkzeug==2.3.6
zipp==3.16.2
azure-functions